UAF-2778 - Promoting delivery fields to PO base class so that PO and 8 subclasses have access. Updating UA extended classes, Spring config, and OJB to reflect new placement. Also correcting misspelling of 'cancelled', since values correspond to preexisting DB data.
