+++ kfs-purap/src/main/java/org/kuali/kfs/module/purap/document/PurchasingDocumentBase.java
@@ -18,11 +18,7 @@
  */
 package org.kuali.kfs.module.purap.document;
 
-import java.sql.Date;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
+import edu.arizona.kfs.sys.businessobject.BuildingExtension;
 import org.apache.commons.lang.StringUtils;
 import org.kuali.kfs.coa.businessobject.Account;
 import org.kuali.kfs.coa.businessobject.Chart;
@@ -51,6 +47,7 @@ import org.kuali.kfs.module.purap.document.service.PurchasingService;
 import org.kuali.kfs.module.purap.document.service.ReceivingAddressService;
 import org.kuali.kfs.module.purap.util.ItemParser;
 import org.kuali.kfs.module.purap.util.ItemParserBase;
+import org.kuali.kfs.sys.businessobject.Building;
 import org.kuali.kfs.sys.context.SpringContext;
 import org.kuali.kfs.sys.service.impl.KfsParameterConstants;
 import org.kuali.kfs.vnd.VendorPropertyConstants;
@@ -72,11 +69,18 @@ import org.kuali.rice.krad.util.ObjectUtils;
 import org.kuali.rice.location.api.country.Country;
 import org.kuali.rice.location.api.country.CountryService;
 
+import java.sql.Date;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
 /**
  * Base class for Purchasing Documents.
  */
 public abstract class PurchasingDocumentBase extends PurchasingAccountsPayableDocumentBase implements PurchasingDocument {
     protected static org.apache.log4j.Logger LOG = org.apache.log4j.Logger.getLogger(PurchasingDocumentBase.class);
+
+    private static final String BUILDING_OBJ_REF_KEY = "buildingObj";
 
     // SHARED FIELDS BETWEEN REQUISITION AND PURCHASE ORDER
     protected String documentFundingSourceCode;
@@ -146,6 +150,8 @@ public abstract class PurchasingDocumentBase extends PurchasingAccountsPayableDo
     protected String capitalAssetSystemTypeCode;
     protected String capitalAssetSystemStateCode;
     protected String justification;
+    protected String routeCode;
+    protected Building buildingObj;
 
     // NOT PERSISTED IN DB
     protected String supplierDiversityLabel;
@@ -1571,5 +1577,32 @@ public abstract class PurchasingDocumentBase extends PurchasingAccountsPayableDo
             return SpringContext.getBean(VendorService.class).getVendorB2BContract(vendorDetail, campusCode) != null;
         }
         return false;
+    }
+
+    public Building getBuildingObj() {
+        if(ObjectUtils.isNull(buildingObj)) {
+            this.refreshReferenceObject(BUILDING_OBJ_REF_KEY);
+        }
+        return buildingObj;
+    }
+
+    public void setBuildingObj(Building buildingObj) {
+        this.buildingObj = buildingObj;
+    }
+
+    public String getRouteCode() {
+        if (StringUtils.isBlank(routeCode)) {
+            try{
+                BuildingExtension be = (BuildingExtension)(getBuildingObj().getExtension());
+                routeCode = be.getRouteCode();
+            }catch(Exception e){
+                LOG.debug("Routing Code was not Retrieved");
+            }
+        }
+        return routeCode;
+    }
+
+    public void setRouteCode(String routeCode) {
+        this.routeCode = routeCode;
     }
 }
