+++ kfs-purap/src/main/resources/edu/arizona/kfs/module/purap/document/datadictionary/PurchaseOrderDocument.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?><beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p" xmlns:dd="http://rice.kuali.org/dd" xsi:schemaLocation="http://www.springframework.org/schema/beans         http://www.springframework.org/schema/beans/spring-beans-2.0.xsd         http://rice.kuali.org/dd         http://rice.kuali.org/dd/dd.xsd">
 
   <bean id="PurchaseOrderDocument" parent="PurchaseOrderDocument-parentBean">
-    <property name="documentClass" value="edu.arizona.kfs.module.purap.document.PurchaseOrderDocument"/>
-    <property name="baseDocumentClass" value="org.kuali.kfs.module.purap.document.PurchaseOrderDocument"/>
+    <property name="documentClass" value="org.kuali.kfs.module.purap.document.PurchaseOrderDocument"/>
     <property name="attributes">
       <list merge="true">
         <ref bean="PurchaseOrderDocument-routeCode"/>
